Always make the CLI Keychain cache available
It doesn't conflict with direct keychain anymore, so no reason to not just make
it always available. This also gives apps the option to use it over direct
access, even if that is available.
