+++ clitoken/cache_test.go
@@ -10,6 +10,22 @@ import (
 	"golang.org/x/oauth2"
 )
 
+func TestKeychainCLICredentialCache(t *testing.T) {
+	// This test requires access to macOS Keychain
+	if os.Getenv("TEST_KEYCHAIN_CLI_CREDENTIAL_CACHE") == "" {
+		t.Skip("TEST_KEYCHAIN_CLI_CREDENTIAL_CACHE not set")
+		return
+	}
+
+	cache := &KeychainCLICredentialCache{}
+
+	if !cache.Available() {
+		t.Fatal("cache is not available")
+	}
+
+	testCache(t, cache)
+}
+
 func TestEncryptedFileCredentialCache(t *testing.T) {
 	dir, err := os.MkdirTemp("", "cachetest")
 	if err != nil {
